#5 implemented "set Movies for Director"

+++ src/main/java/movie/core/repository/DirectorRepository.java
@@ -3,7 +3,7 @@ package movie.core.repository;
 import movie.core.model.Director;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DirectorRepository extends CrudRepository<Director, Integer> {
+public interface DirectorRepository extends CrudRepository<Director, Long> {
 
 
 }
